pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "csv-stream-diff"
version = "0.2.0"
description = "Stream and compare very large CSV files with multiprocessing."
authors = ["Jordi"]
license = "MIT"
readme = "README.md"
keywords = ["csv", "diff", "streaming", "multiprocessing", "comparison"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: File Formats",
  "Topic :: Software Development :: Testing",
  "Topic :: Utilities",
]
packages = [
  { include = "csvstreamdiff", from = "src" },
]
include = [
  { path = "config.example.yaml", format = ["sdist", "wheel"] },
  { path = "BLOGPOST.md", format = "sdist" },
  { path = "generator", format = "sdist" },
  { path = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
PyYAML = ">=6.0"
rich = ">=13.7"

[tool.poetry.group.dev.dependencies]
behave = ">=1.2.6"
pytest = ">=7.4"

[tool.poetry.scripts]
csv-stream-diff = "csvstreamdiff.cli:main"
